DQNAgent in separate file
